requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   groq
    #   httpx
    #   sse-starlette
    #   starlette
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via uvicorn
colorama==0.4.6
    # via
    #   click
    #   tqdm
contourpy==1.3.2
    # via matplotlib
cycler==0.12.1
    # via matplotlib
deep-sort-realtime==1.3.2
    # via backend (pyproject.toml)
distro==1.9.0
    # via groq
fastapi==0.115.12
    # via backend (pyproject.toml)
filelock==3.18.0
    # via torch
fonttools==4.57.0
    # via matplotlib
fsspec==2025.3.2
    # via torch
greenlet==3.2.1
    # via sqlalchemy
groq==0.23.0
    # via langchain-groq
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.8
    # via httpx
httpx==0.28.1
    # via
    #   groq
    #   langsmith
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.6
    # via torch
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
kiwisolver==1.4.8
    # via matplotlib
langchain==0.3.24
    # via backend (pyproject.toml)
langchain-core==0.3.55
    # via
    #   langchain
    #   langchain-groq
    #   langchain-text-splitters
langchain-groq==0.3.2
    # via backend (pyproject.toml)
langchain-text-splitters==0.3.8
    # via langchain
langsmith==0.3.33
    # via
    #   langchain
    #   langchain-core
markupsafe==3.0.2
    # via jinja2
matplotlib==3.10.1
    # via
    #   seaborn
    #   ultralytics
mpmath==1.3.0
    # via sympy
networkx==3.4.2
    # via torch
numpy==2.1.1
    # via
    #   backend (pyproject.toml)
    #   contourpy
    #   deep-sort-realtime
    #   matplotlib
    #   opencv-python
    #   pandas
    #   scipy
    #   seaborn
    #   torchvision
    #   ultralytics
    #   ultralytics-thop
opencv-python==4.11.0.86
    # via
    #   backend (pyproject.toml)
    #   deep-sort-realtime
    #   ultralytics
orjson==3.10.16
    # via langsmith
packaging==24.2
    # via
    #   langchain-core
    #   langsmith
    #   matplotlib
pandas==2.2.3
    # via
    #   seaborn
    #   ultralytics
pillow==11.2.1
    # via
    #   backend (pyproject.toml)
    #   matplotlib
    #   torchvision
    #   ultralytics
psutil==7.0.0
    # via ultralytics
py-cpuinfo==9.0.0
    # via ultralytics
pydantic==2.11.3
    # via
    #   fastapi
    #   groq
    #   langchain
    #   langchain-core
    #   langsmith
pydantic-core==2.33.1
    # via pydantic
pyparsing==3.2.3
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
python-dotenv==1.1.0
    # via backend (pyproject.toml)
python-multipart==0.0.20
    # via backend (pyproject.toml)
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   langchain
    #   langchain-core
    #   ultralytics
requests==2.32.3
    # via
    #   langchain
    #   langsmith
    #   requests-toolbelt
    #   ultralytics
requests-toolbelt==1.0.0
    # via langsmith
scipy==1.15.2
    # via
    #   deep-sort-realtime
    #   ultralytics
seaborn==0.13.2
    # via ultralytics
setuptools==79.0.1
    # via backend (pyproject.toml)
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   groq
sqlalchemy==2.0.40
    # via langchain
sse-starlette==2.3.3
    # via backend (pyproject.toml)
starlette==0.46.2
    # via
    #   fastapi
    #   sse-starlette
sympy==1.13.3
    # via torch
tenacity==9.1.2
    # via langchain-core
torch==2.7.0
    # via
    #   backend (pyproject.toml)
    #   torchvision
    #   ultralytics
    #   ultralytics-thop
torchvision==0.22.0
    # via ultralytics
tqdm==4.67.1
    # via ultralytics
typing-extensions==4.13.2
    # via
    #   anyio
    #   fastapi
    #   groq
    #   langchain-core
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   torch
    #   typing-inspection
typing-inspection==0.4.0
    # via pydantic
tzdata==2025.2
    # via pandas
ultralytics==8.3.115
    # via backend (pyproject.toml)
ultralytics-thop==2.0.14
    # via ultralytics
urllib3==2.4.0
    # via requests
uvicorn==0.34.2
    # via backend (pyproject.toml)
zstandard==0.23.0
    # via langsmith
